+++ Runtime/Module/Status/StatusModule.cs
@@ -5,12 +5,14 @@ using AffiseAttributionLib.Executors;
 using AffiseAttributionLib.Module.Status.UseCase;
 using AffiseAttributionLib.Modules;
 using AffiseAttributionLib.Network;
+using AffiseAttributionLib.Referrer;
 
 namespace AffiseAttributionLib.Module.Status
 {
-    public class StatusModule : AffiseModule
+    public class StatusModule : AffiseModule, ReferrerCallback
     {
         private ICheckStatusUseCase? _checkStatusUseCase;
+        private IReferrerUseCase? _referrerUseCase;
 
         public override void Start()
         {
@@ -27,18 +29,31 @@ namespace AffiseAttributionLib.Module.Status
                 providersToJsonConverter,
                 new StringToKeyValueConverter()
             );
+
+            _referrerUseCase = new ReferrerUseCaseImpl(
+                _checkStatusUseCase
+            );
         }
 
         public override void Status(OnKeyValueCallback onComplete)
         {
             if (_checkStatusUseCase is not null)
             {
-                _checkStatusUseCase?.Send(onComplete);
+                _checkStatusUseCase?.Send((data) =>
+                {
+                    onComplete.Invoke(data);
+                    _referrerUseCase?.ParseStatus(data);
+                });
             }
             else
             {
                 onComplete(new List<AffiseKeyValue>());
             }
+        }
+
+        public void GetReferrer(OnReferrerCallback callback)
+        {
+            _referrerUseCase?.GetReferrer(callback);
         }
     }
 }