+++ Runtime/Referrer/OnReferrerCallback.cs
@@ -1,0 +1,5 @@
+﻿#nullable enable
+namespace AffiseAttributionLib.Referrer
+{
+    public delegate void OnReferrerCallback(string? value);
+}