+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.6.30] - 2025-01-29
+
+### Added
+
+- Api `Affise.GetReferrerOnServer`.
+- Api `Affise.GetReferrerOnServerValue`.
+
+### Changed
+
+- Update native Android to [`1.6.51`](https://github.com/affise/sdk-android/blob/v1.6.51/CHANGELOG.md).
+
 ## [1.6.29] - 2024-12-20
 
 ### Changed
@@ -189,6 +200,7 @@
 - Update native iOS to `1.6.20`.
 - Update native Android to `1.6.22`.
   
+[1.6.30]: https://github.com/affise/sdk-unity/compare/1.6.29...1.6.30
 [1.6.29]: https://github.com/affise/sdk-unity/compare/1.6.28...1.6.29
 [1.6.28]: https://github.com/affise/sdk-unity/compare/1.6.27...1.6.28
 [1.6.27]: https://github.com/affise/sdk-unity/compare/1.6.26...1.6.27